Handle YouTube encode links, optimize getFiles() via recusion
If a ghost does not exist in the repo, replace link with a YouTube encode of the  run.

getFiles() now uses recursion to navigate down a series of folders.

Print-out for 3lap/flap times is  now identical (includes 3lap and individual lap times) which results in a bit of cleanup